Fix: Alt+F4/Windows-luk på progress-dialogen cancellede ikke testen
Code review fandt at kun Stop-knappen og Esc udløste cancellation via
Cancel_Click — lukkes vinduet på anden vis (Alt+F4, titlebar-luk hvis den
nogen sinde vises, programmatisk Close()), fortsatte ping-løkken usynligt
i op til testens fulde varighed. Samme "frossen"-oplevelse som den
oprindelige bug, blot via en anden indgang.

Flyttet cancellation-logikken til en Closing-handler, som fanger enhver
lukkevej ensartet. Cancel_Click er nu blot en tynd Close()-wrapper.

Fjernet også et par overflødige (int)-casts i PingMonitorViewModel —
ConnectionTestStats.Sent/Replies er allerede int.

+++ ViewModels/PingMonitorViewModel.cs
@@ -195,8 +195,8 @@ namespace M1Scan.ViewModels
                     if (Application.Current != null)
                         await Application.Current.Dispatcher.InvokeAsync(() =>
                             progressDialog.UpdateProgress(elapsedSeconds,
-                                (int)targetStats.Sent, (int)targetStats.Replies,
-                                (int)referenceStats.Sent, (int)referenceStats.Replies));
+                                targetStats.Sent, targetStats.Replies,
+                                referenceStats.Sent, referenceStats.Replies));
 
                     var remaining = nextTick - DateTime.UtcNow;
                     if (remaining > TimeSpan.Zero)
